+++ HideousGameEngine/test/RectColorTest/RunningSquare.cpp
@@ -27,7 +27,7 @@ sprite_(0),
 life_state_(kLiving),
 color_(GLKVector4Make(he::Randf(), he::Randf(), he::Randf(),he::Randf()))
 {
-	double dimension = he::Randf() * 40.0 + 2.0;
+	float dimension = he::Randf() * 40.0 + 2.0;
 	vertex_data_ = new he::VertexCol(dimension/2, dimension/2);
 	sprite_ = new he::RenderObject(vertex_data_, shader, 0, he::g_Screen->projection_, color_);
 	transform_.position_ = GLKVector2Make(-he::g_Screen->width_/2+he::Randf()*he::g_Screen->width_, -he::g_Screen->height_/2+he::Randf()*he::g_Screen->height_);
@@ -41,7 +41,7 @@ void RunningSquare::SetDead(){
 	life_state_ = kDead;
 }
 
-void RunningSquare::Update(double dt){
+void RunningSquare::Update(float dt){
 
 	if(life_state_ != kLiving)
 		return;
