+++ HideousGameEngine/test/RectTextTest/RectTextTest.h
@@ -20,8 +20,8 @@ namespace he{
 class RectTextTest : public ITest{
 public:
 	~RectTextTest();
-	RectTextTest(double width, double height);
-	void Update(double dt);
+	RectTextTest(float width, float height);
+	void Update(float dt);
 	void Render();
 	void HandleGesture(const he::Gesture &gesture);
