+++ HideousGameEngine/test/TestFactory.h
@@ -11,57 +11,57 @@
 
 #include "RectColorTest/RectColorTest.h"
 //Test cases factories
-RectColorTest *MakeRectColorTest(double w, double h){
+RectColorTest *MakeRectColorTest(float w, float h){
 	return new RectColorTest(w, h);
 }
 
 #include "RectTextureTest/RectTextureTest.h"
-RectTextureTest *MakeRectTextureTest(double w, double h){
+RectTextureTest *MakeRectTextureTest(float w, float h){
 	return new RectTextureTest(w, h);
 }
 
 #include "RectTextTest/RectTextTest.h"
-RectTextTest *MakeRectTextTest(double w, double h){
+RectTextTest *MakeRectTextTest(float w, float h){
 	return new RectTextTest(w, h);
 }
 
 #include "MultiShadersTest/MultiShadersTest.h"
-MultiShadersTest *MakeMultiShadersTest(double w, double h){
+MultiShadersTest *MakeMultiShadersTest(float w, float h){
 	return new MultiShadersTest(w, h);
 }
 
 #include "MultiFontTest/MultiFontTest.h"
-MultiFontTest *MakeMultiFontTest(double w, double h){
+MultiFontTest *MakeMultiFontTest(float w, float h){
 	return new MultiFontTest(w,h);
 }
 
 #include "AnimationTest/AnimationTest.h"
-AnimationTest *MakeAnimationTest(double w, double h){
+AnimationTest *MakeAnimationTest(float w, float h){
 	return new AnimationTest(w, h);
 }
 
 #include "TextureBigBangTest/TextureBigBangTest.h"
-TextureBigBangTest *MakeTextureBigBangTest(double w, double h){
+TextureBigBangTest *MakeTextureBigBangTest(float w, float h){
 	return new TextureBigBangTest(w,h);
 }
 
 #include "GestureTest/GestureTest.h"
-GestureTest *MakeGestureTest(double w, double h){
+GestureTest *MakeGestureTest(float w, float h){
 	return new GestureTest(w,h);
 }
 
 #include "UITest/UITest.h"
-UITest *MakeUITest(double w, double h){
+UITest *MakeUITest(float w, float h){
 	return new UITest(w,h);
 }
 
 #include "ParticleTest/ParticleTest.h"
-ParticleTest *MakeParticlesTest(double w, double h){
+ParticleTest *MakeParticlesTest(float w, float h){
 	return new ParticleTest(w,h);
 }
 
 #include "SpineTest/SpineTest.h"
-SpineTest *MakeSpineTest(double w, double h){
+SpineTest *MakeSpineTest(float w, float h){
 	return new SpineTest(w, h);
 }
 #endif
